CORE: save times in sql as 32 bit for now

+++ src/map/linkshell.cpp
@@ -85,8 +85,8 @@ void CLinkshell::setMessage(const int8* message, const int8* poster)
 {
     int8 sqlMessage[256];
     Sql_EscapeString(SqlHandle, sqlMessage, message);
-    Sql_Query(SqlHandle, "UPDATE linkshells SET poster = '%s', message = '%s', messagetime = %llu WHERE linkshellid = %d;",
-        poster, sqlMessage , time(nullptr), m_id);
+    Sql_Query(SqlHandle, "UPDATE linkshells SET poster = '%s', message = '%s', messagetime = %u WHERE linkshellid = %d;",
+        poster, sqlMessage , static_cast<uint32>(time(nullptr)), m_id);
 
     int8 packetData[8] {};
     WBUFL(packetData, 0) = m_id;
